requirements.txt
acme==1.21.0
anyio==4.9.0
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certbot==1.21.0
certbot-nginx==1.21.0
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloud-init==24.4.1
colorama==0.4.4
command-not-found==0.3
ConfigArgParse==1.5.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.3.0
filelock==3.18.0
gunicorn==23.0.0
h11==0.16.0
httplib2==0.20.2
httptools==0.6.4
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
josepy==1.10.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==21.3
parsedatetime==2.6
pexpect==4.8.0
platformdirs==4.3.7
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyICU==2.8.1
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.1.0
python-magic==0.4.24
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.9.1
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sniffio==1.3.1
sos==4.8.2
ssh-import-id==5.11
systemd-python==234
Twisted==22.1.0
typing_extensions==4.13.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
uvicorn==0.34.2
uvloop==0.21.0
virtualenv==20.31.0
wadllib==1.3.6
watchfiles==1.0.5
websockets==15.0.1
xkit==0.0.0
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0
